rename new comment related code
