+++ components/screens/post/NewCommentScreen.tsx
@@ -6,9 +6,9 @@ import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { Ionicons } from "@expo/vector-icons";
 import moment from "moment/moment";
 import { useAppSelector } from "../../../store";
-import { selectNewComment } from "../../../slices/newComment/newCommentSlice";
+import { selectNewComment } from "../../../slices/comments/newCommentSlice";
 import LoadingView from "../../ui/Loading/LoadingView";
-import useNewComment from "../../hooks/newComment/newCommentHooks";
+import useNewComment from "../../hooks/comments/useNewComment";
 import { truncateName } from "../../../helpers/TextHelper";
 import RenderMarkdown from "../../ui/markdown/RenderMarkdown";
 import KeyboardAccessory from "../../ui/KeyboardAccessory";
