+++ components/ui/Feed/CompactFeedItem/CompactFeedItem.tsx
@@ -14,7 +14,7 @@ import Animated from "react-native-reanimated";
 import useFeedItem from "../../../hooks/feeds/useFeedItem";
 import { ILemmyVote } from "../../../../lemmy/types/ILemmyVote";
 import useSwipeAnimation from "../../../hooks/animations/useSwipeAnimation";
-import { setResponseTo } from "../../../../slices/newComment/newCommentSlice";
+import { setResponseTo } from "../../../../slices/comments/newCommentSlice";
 import { useAppDispatch, useAppSelector } from "../../../../store";
 import CompactFeedItemThumbnail from "./CompactFeedItemThumbnail";
 import CompactFeedItemVote from "./CompactFeedItemVote";
