+++ slices/post/postActions.ts
@@ -1,8 +1,8 @@
-import {createAsyncThunk} from "@reduxjs/toolkit";
-import {GetPost} from "lemmy-js-client";
-import {lemmyInstance} from "../../lemmy/LemmyInstance";
-import {setPost} from "./postSlice";
-import {writeToLog} from "../../helpers/LogHelper";
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import { GetPost } from "lemmy-js-client";
+import { lemmyInstance } from "../../lemmy/LemmyInstance";
+import { setPost } from "./postSlice";
+import { writeToLog } from "../../helpers/LogHelper";
 
 // eslint-disable-next-line import/prefer-default-export
 export const getPost = createAsyncThunk(
