+++ store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import { postSlice } from "./slices/post/postSlice";
-import { newCommentSlice } from "./slices/newComment/newCommentSlice";
+import { newCommentSlice } from "./slices/comments/newCommentSlice";
 import feedSlice from "./slices/feed/feedSlice";
 import communitiesSlice from "./slices/communities/communitiesSlice";
 import settingsSlice from "./slices/settings/settingsSlice";
